refactor: Functions, Extensions 분류 뒤 카테고리 주석 추가
