+++ app/src/main/java/com/treasurehunt/util/Extensions.kt
@@ -2,13 +2,11 @@ package com.treasurehunt.util
 
 import android.view.View
 import com.google.android.material.snackbar.Snackbar
-import com.naver.maps.geometry.LatLng
 import kotlinx.serialization.json.Json
 import java.math.RoundingMode
 import java.text.DecimalFormat
-import kotlin.math.pow
 
-//******************** View ********************
+/* ----------- View ----------- */
 
 fun View.showSnackbar(resId: Int) {
     Snackbar.make(
@@ -34,10 +32,9 @@ fun View.hide() {
     visibility = View.GONE
 }
 
+/* ----------- String ----------- */
 
-//******************** String ********************
-
-internal inline fun <reified R> String.convertToDataClass() = json.decodeFromString<R>(this)
+internal fun String.extractDigits() = replace("[^0-9]".toRegex(), "")
 
 private val json = Json {
     isLenient = true
@@ -45,29 +42,9 @@ private val json = Json {
     coerceInputValues = true
 }
 
-internal fun String.extractDigits() = replace("[^0-9]".toRegex(), "")
+internal inline fun <reified R> String.convertToDataClass() = json.decodeFromString<R>(this)
 
-private val coordMultiplier = 10.0.pow(7)
-
-// currently, Naver Local Search API mapx, mapy implement WGS84 (LatLng) coord system as integer
-// beware that y, x correspond to lat, lng respectively
-fun convertMapXYToLatLng(xy: Pair<String?, String?>): LatLng? {
-    val x = xy.first?.toLongOrNull() ?: return null
-    val y = xy.second?.toLongOrNull() ?: return null
-    return LatLng(y / coordMultiplier, x / coordMultiplier)
-}
-
-private const val KILOMETER_BREAKPOINT = 1000.0
-private const val UNIT_METER = "m"
-private const val UNIT_KILOMETER = "km"
-
-fun formatDistance(meter: Long): String {
-    require(meter >= 0)
-
-    if (meter < KILOMETER_BREAKPOINT) return "${meter}$UNIT_METER"
-
-    return "${(meter / KILOMETER_BREAKPOINT).roundOff()}$UNIT_KILOMETER"
-}
+/* ----------- Double ----------- */
 
 fun Double.roundOff(): Double = DecimalFormat("#.##").apply {
     roundingMode = RoundingMode.CEILING
