+++ app/src/main/java/com/treasurehunt/util/Functions.kt
@@ -5,6 +5,7 @@ import com.google.firebase.Firebase
 import com.google.firebase.FirebaseException
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.auth
+import com.naver.maps.geometry.LatLng
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.channels.trySendBlocking
 import kotlinx.coroutines.flow.Flow
@@ -12,6 +13,9 @@ import kotlinx.coroutines.flow.callbackFlow
 import java.time.LocalDateTime
 import java.time.ZoneId
 import java.util.Date
+import kotlin.math.pow
+
+/* ----------- 시간 ----------- */
 
 fun getCurrentTime(): Long = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
     LocalDateTime.now().atZone(ZoneId.systemDefault()).toInstant().toEpochMilli()
@@ -34,4 +38,32 @@ internal fun getUidCallbackFlow(): Flow<String?> = callbackFlow {
     awaitClose {
         Firebase.auth.removeAuthStateListener(callback)
     }
+}
+
+/* ----------- 좌표 ----------- */
+
+private val coordMultiplier = 10.0.pow(7)
+
+/**
+ * currently, Naver Local Search API mapx, mapy implement WGS84 (LatLng) coord system as integer
+ * beware that y, x correspond to lat, lng respectively
+ * */
+fun convertMapXYToLatLng(xy: Pair<String?, String?>): LatLng? {
+    val x = xy.first?.toLongOrNull() ?: return null
+    val y = xy.second?.toLongOrNull() ?: return null
+    return LatLng(y / coordMultiplier, x / coordMultiplier)
+}
+
+/* ----------- 거리 ----------- */
+
+private const val KILOMETER_BREAKPOINT = 1000.0
+private const val UNIT_METER = "m"
+private const val UNIT_KILOMETER = "km"
+
+fun formatDistance(meter: Long): String {
+    require(meter >= 0)
+
+    if (meter < KILOMETER_BREAKPOINT) return "${meter}$UNIT_METER"
+
+    return "${(meter / KILOMETER_BREAKPOINT).roundOff()}$UNIT_KILOMETER"
 }
